+++ CHANGELOG.md
@@ -4,6 +4,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0] - 2020-08-31
+### Added
+- AWS CDK and AWS Solutions Constructs to create AWS CloudFormation template
+
+### Fixed
+- Auto WebP does not work properly: [#195](https://github.com/awslabs/serverless-image-handler/pull/195), [#200](https://github.com/awslabs/serverless-image-handler/issues/200), [#205](https://github.com/awslabs/serverless-image-handler/issues/205)
+- A bug where base64 encoding containing slash: [#194](https://github.com/awslabs/serverless-image-handler/pull/194)
+- Thumbor issues:
+  - `0` size support: [#183](https://github.com/awslabs/serverless-image-handler/issues/183)
+  - `convolution` filter does not work: [#187](https://github.com/awslabs/serverless-image-handler/issues/187)
+  - `fill` filter does not work: [#190](https://github.com/awslabs/serverless-image-handler/issues/190)
+- __Note that__ duplicated features has been merged gracefully.
+
+### Removed
+- AWS CloudFormation template: `serverless-image-handler.template`
+
+### Changed
+- sharp base version (from 0.23.4 to 0.25.4)
+- Remove `Promise` to return since `async` functions return promises: [#189](https://github.com/awslabs/serverless-image-handler/issues/189)
+- Unit test statement coverage improvement:
+  - `image-handler.js`: `79.05%` to `100%`
+  - `image-request.js`: `93.58%` to `100%`
+  - `thumbor-mapping.js`: `99.29%` to `100%`
+  - `overall`: `91.55%` to `100%`
+
 ## [4.2] - 2020-02-06
 ### Added
 - Honor outputFormat Parameter from the pull request [#117](https://github.com/awslabs/serverless-image-handler/pull/117)
@@ -16,7 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix: DeprecationWarning: Buffer() is deprecated from the pull request [#174](https://github.com/awslabs/serverless-image-handler/pull/174)
 - Add hex color support for Thumbor ```filters:background_color``` and ```filters:fill``` [#154](https://github.com/awslabs/serverless-image-handler/issues/154)
 - Add format and watermark support for Thumbor [#109](https://github.com/awslabs/serverless-image-handler/issues/109), [#131](https://github.com/awslabs/serverless-image-handler/issues/131), [#109](https://github.com/awslabs/serverless-image-handler/issues/142)
-* __Note that__ duplicated features has been merged gracefully.
+- __Note that__ duplicated features has been merged gracefully.
 
 ### Changed
 - sharp base version (from 0.23.3 to 0.23.4)